backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic-settings
langchain-sqlserver==0.1.1
langchain-huggingface
sentence-transformers
langchain-community
faiss-cpu
langchain-openai
pymupdf4llm
python-multipart
sqlalchemy
pyodbc
moviepy==1.0.3
pydub
aiofiles